requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
stripe>=7.0.0
requests>=2.31.0
pillow>=10.0.0
aiofiles>=23.0.0
python-dotenv>=1.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
python-jose[cryptography]>=3.3.0
cryptography>=40.0.0
psycopg2-binary>=2.9.0